test/python/requirements.txt
# Python dependencies for BM25 validation system
psycopg2-binary>=2.9.0
rank-bm25>=0.2.0
